backend/requirements.txt
annotated-types==0.7.0
anthropic==0.49.0
anyio==4.9.0
blessed==1.20.0
boto3==1.37.23
botocore==1.37.23
cachetools==5.5.2
certifi==2025.1.31
chalice==1.31.4
charset-normalizer==3.4.1
click==8.1.8
distro==1.9.0
docstring_parser==0.16
editor==1.6.6
google-api-core==2.24.2
google-auth==2.38.0
google-cloud-aiplatform==1.71.1
google-cloud-bigquery==3.31.0
google-cloud-core==2.4.3
google-cloud-resource-manager==1.14.2
google-cloud-storage==2.19.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.69.2
grpc-google-iam-v1==0.14.2
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
inquirer==3.4.0
jiter==0.9.0
jmespath==1.0.1
numpy==2.2.4
openai==1.69.0
packaging==24.2
proto-plus==1.26.1
protobuf==5.29.4
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.1
pydantic_core==2.33.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
PyYAML==6.0.2
readchar==4.2.1
requests==2.32.3
rsa==4.9
runs==1.2.2
s3transfer==0.11.4
setuptools==78.1.0
shapely==2.1.0rc1
six==1.17.0
sniffio==1.3.1
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.0
urllib3==2.3.0
wcwidth==0.2.13
wheel==0.45.1
xmod==1.8.1
